Make sure to trigger area exit and queue logout on server shutdown
